reversionando la clase modelo controler agregando validaciones a conexion a base de datos y estilos

+++ modelo/index.php
@@ -1,30 +1,39 @@
 <?php
-class Modelo { //es generico puedo utilizarlo para cualquier tabla de la base de datos mvc
+class Modelo { //es generico puedo utilizarlo para cualquier tabla de la base de datos mvc.
     
-    private $Modelo;
+    
     private $db;
     private $datos;
     
-    //metodo constructor
+    //metodo constructor.
     public function __construct(){
-        $this->Modelo = array();
-        $this->db = new PDO('mysql:host=localhost;dbname=mvc',"root","");
+        $this->datos = [];
+    }
+
+    //funcion para conectar a la base de datos.
+    public function conectar() {
+        try {
+            $this->db = new PDO('mysql:host=localhost;dbname=mvc', 'root', '');
+            $this->db->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+            return true;
+        } catch (PDOException $e) {
+            return $e->getMessage(); // Retorno el error para manejarlo en el controlador.
+        }
     }
 
     //funcion para insertar datos a la tabla
     public function insertar($tabla,$data){
+        if (!$this->db) return false;
         $consulta = "insert into ".$tabla." values(null,".$data.")";
         $resultado=$this->db->query($consulta);
-        if($resultado){
-            return true;
-        }else{
-            return false;
-        }
+        return $resultado ? true : false;
     }
 
     //funcion para mostrar datos de la tabla
     public function mostrar($tabla,$condicion){
+        if (!$this->db) return []; // no hacer nada si no hay conexión
         $consul="select * from ".$tabla." where ".$condicion.";";
+        
         $resu=$this->db->query($consul);
         while($filas = $resu->FETCHALL(PDO::FETCH_ASSOC)) {
             $this->datos[]=$filas;
@@ -34,23 +43,17 @@ class Modelo { //es generico puedo utilizarlo para cualquier tabla de la base de
 
     //funcion para actualizar la tabla
     function actualizar($tabla,$data,$condicion){
+        if (!$this->db) return false;
         $consulta="update ".$tabla." set ".$data." where ".$condicion;
         $resultado=$this->db->query($consulta);
-        if($resultado){
-            return true;
-        }else{
-            return false;
-        }
+        return $resultado ? true : false;
     }
 
     //funcion para eliminar registro de una tabla
     function eliminar($tabla,$condicion){
+        if (!$this->db) return false;
         $eli="delete from ".$tabla." where ".$condicion;
         $res=$this->db->query($eli);
-        if($res){
-            return true;
-        }else{
-            return false;
-        }
+        return $res ? true : false;
     }
 }
